Feat/buffer classes and refactoring (#116)
* refactor: changed AudioTrack buffer type to FloatArray and refactored channelCount prop

* feat: implemented ChannelCountMode and ChannelInterpretation

* refactor: object construction refactoring

* refactor: yarn format

* refactor: changed folder name and removed unused files

* refactor: general ios classes refactoring

* feat: implemented AudioBuffer

* feat: implemented AudioBuffer

* refactor: yarn format

* fix: marked AudioBufferWrapper constructor for IOS as explicit

---------

+++ android/src/main/cpp/AudioBuffer/AudioBuffer.h
@@ -32,8 +32,8 @@ class AudioBuffer : public jni::HybridClass<AudioBuffer> {
   int getLength();
   double getDuration();
   int getNumberOfChannels();
-  int16_t **getChannelData(int channel);
-  void setChannelData(int channel, int16_t **data);
+  float **getChannelData(int channel);
+  void setChannelData(int channel, float **data);
   void prepareForDeconstruction();
 
  public:
